# src/main/resources/static/api-docs.yaml
openapi: 3.1.0
info:
  title: SocialMedia OpenApi Definition
  description: Open Api for SocialMedia project
  version: '1'
servers:
  - url: http://127.0.0.1:8081/api
    description: Generated server url
...
paths:
...
  /blacklist:
    get:
      tags:
        - Black List
      summary: Get blacklist of an Authenticated user
      operationId: allBlacklistedUsers
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
                uniqueItems: true
      security:
        - bearerAuth: []
    post:
      tags:
        - Black List
      summary: Add user (id) to blacklist
      operationId: addToBlackList
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: added!
        '400':
          description: BAD_REQUEST
          content:
            text/plain:
              schema:
                type: string
                example: This is you!
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: User not found!
      security:
        - bearerAuth: []
    delete:
      tags:
        - Black List
      summary: Remove user(id) from blacklist
      operationId: removeFromBlackList
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: removed!
        '400':
          description: BAD_REQUEST
          content:
            text/plain:
              schema:
                type: string
                example: User is not in blacklist
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: User not found!
      security:
        - bearerAuth: []
  /a/register:
    post:
      tags:
        - Authentication
      summary: Sign Up in the System
      description: Create a new account in the System
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: Bearer token for authentication
        '400':
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserErrorResponse'
  /a/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh expired token
      description: Refresh expired Jwt token via Refresh token from cookie
      operationId: refreshToken
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: Bearer token for authentication
        '401':
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserErrorResponse'
  /a/login:
    post:
      tags:
        - Authentication
      summary: Sign In to the System
      description: Get a Bearer token via Authentication
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: Bearer token for authentication
        '401':
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserErrorResponse'
...
  /admin/users/unban/{id}:
    patch:
      tags:
        - Admin Functions
      summary: Unblock user
      operationId: unban
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: User 1 is unbanned
        '400':
          description: BAD_REQUEST
          content:
            text/plain:
              schema:
                type: string
                example: This user is not banned
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: User not found!
      security:
        - bearerAuth: []
  /admin/users/ban/{id}:
    patch:
      tags:
        - Admin Functions
      summary: Block user
      description: Block user (id) if the user has violated the rules.
      operationId: ban
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: User 1 is banned
        '400':
          description: BAD_REQUEST
          content:
            text/plain:
              schema:
                type: string
                example: This user is already banned
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: User not found!
      security:
        - bearerAuth: []
  /admin/assign/{id}:
    patch:
      tags:
        - Admin Functions
      summary: Assign user as administrator
      operationId: assign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: User 1 added to admins
        '400':
          description: BAD_REQUEST
          content:
            text/plain:
              schema:
                type: string
                example: This is you!
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: User not found!
      security:
        - bearerAuth: []
...
  /posts/{id}:
    get:
      tags:
        - Posts
      summary: Get post by id
      operationId: post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PostResponseDTO'
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Post not found!
...
  /admin/check:
    get:
      tags:
        - Admin Functions
      summary: Check administrator access
      description: Returns 200(OK) if user is administrator, or 401(UNAUTHORIZED) if user isn't administrator
      operationId: accessGranted
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
                example: true
        '401':
          description: UNAUTHORIZED
      security:
        - bearerAuth: []
...
  /admin/posts/{id}:
    delete:
      tags:
        - Admin Functions
      summary: Delete post
      description: Delete post (id) if it has violated the rules.
      operationId: deletePost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: Post 1 is deleted
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Post not found!
      security:
        - bearerAuth: []
components:
  schemas:
    UserErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Error message
...
    RegisterRequestDTO:
      type: object
      properties:
        email:
          type: string
          example: test@test.com
        password:
          type: string
          example: 12345678
    LoginRequestDTO:
      type: object
      properties:
        email:
          type: string
          example: test@test.com
        password:
          type: string
          example: 12345678
      required:
        - email
        - password
...
    PostResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        photoUrl:
          type: string
        text:
          type: string
        countOfLike:
          type: integer
          format: int64
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        dateOfBirth:
          type: string
        profilePicture:
          type: string
        country:
          type: string
        private:
          type: boolean
        valid:
          type: boolean
        blocked:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT